requirements.txt
black==21.12b0
dj-database-url==0.5.0
Django==4.0.1
django-debug-toolbar==3.2.4
environs==9.4.0
flake8==4.0.1
whitenoise==5.3.0
djangorestframework==3.13.1
